move method to utility
